requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
Flask-WTF
WTForms
email-validator
celery
redis
python-dotenv